requirements.txt
Flask
tensorflow>=2.15
Pillow
numpy
